Improve launch_command
